rename renderer to encoder
